+++ test/primitives/null-element-test.js
@@ -23,25 +23,6 @@ describe('NullElement', function() {
     });
   });
 
-  describe('#toValue', function() {
-    it('returns null', function() {
-      expect(nullElement.toValue()).to.equal(null);
-    });
-  });
-
-  describe('#toRefract', function() {
-    var expected = {
-      element: 'null',
-      meta: {},
-      attributes: {},
-      content: null
-    };
-
-    it('returns a null element', function() {
-      expect(nullElement.toRefract()).to.deep.equal(expected);
-    });
-  });
-
   describe('#get', function() {
     it('returns the null value', function() {
       expect(nullElement.toValue()).to.equal(null);
@@ -51,15 +32,6 @@ describe('NullElement', function() {
   describe('#set', function() {
     it('cannot set the value', function() {
       expect(nullElement.set('foobar')).to.be.an.instanceof(Error);
-    });
-  });
-
-  describe('#clone', function() {
-    it('creates a deep clone of the element', function() {
-      var clone = nullElement.clone();
-      expect(clone).to.be.instanceOf(NullElement);
-      expect(clone).to.not.equal(nullElement);
-      expect(clone.toRefract()).to.deep.equal(nullElement.toRefract());
     });
   });
 });